Consolidate test-suite rendering harnesses (#2236)
