+++ .gitmodules
@@ -1,7 +1,3 @@
-[submodule "test/suite"]
-	path = test/suite
-	url = https://github.com/mapbox/mapbox-gl-test-suite.git
-
 [submodule ".mason"]
 	path = .mason
 	url = https://github.com/mapbox/mason.git
