--- scripts/compare_images.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-set -e
-set -o pipefail
-
-(cd ./test/suite/ && ./bin/compare_images.py native)
